requirements.txt
amqp==5.0.7
asgiref==3.4.1
Babel==2.9.1
billiard==3.6.4.0
boto3==1.20.24
botocore==1.23.24
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.9
django-ckeditor==6.1.0
django-crispy-forms==1.13.0
django-js-asset==1.2.2
django-phone-login==1.1.3
django-phonenumber-field==1.1.0
django-sendsms==0.2.3
django-storages==1.12.3
django-utils-six==2.0
djangorestframework==3.12.4
ghasedak==0.1.7
idna==3.3
jmespath==0.10.0
kavenegar==1.1.2
kombu==5.2.2
phonenumberslite==8.12.37
Pillow
prompt-toolkit==3.0.24
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
celery

#oauth
social-auth-app-django

# rest farmework
djangorestframework
djangorestframework-oauth
djangorestframework-simplejwt